refactor: move logic into small files
